api/requirements.txt
requests==2.31.0
django-filter==23.2
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
